requirements.txt
python-telegram-bot
yt-dlp
pycryptodomex
python-dotenv 